fix import paths, remove task modal and create task detail view
