+++ packages/frinx-workflow-ui/src/components/modals/confirm-delete-workflow-modal.tsx
@@ -13,7 +13,7 @@ import {
   Text,
 } from '@chakra-ui/react';
 import FeatherIcon from 'feather-icons-react';
-import { ClientWorkflow } from 'packages/shared/src';
+import { ClientWorkflow } from '@frinx/shared';
 
 type Props = {
   activeWorkflow: ClientWorkflow;
