+++ packages/frinx-workflow-ui/src/components/modals/index.ts
@@ -2,7 +2,6 @@ import DefinitionModal from './definition-modal';
 import DependencyModal from './dependency-modal';
 import DiagramModal from './diagram-modal';
 import ScheduleWorkflowModal from './schedule-workflow-modal';
-import TaskModal from './task-modal';
 import ConfirmDeleteModal from './confirm-delete-workflow-modal';
 
-export { DefinitionModal, DependencyModal, DiagramModal, ScheduleWorkflowModal, TaskModal, ConfirmDeleteModal };
+export { DefinitionModal, DependencyModal, DiagramModal, ScheduleWorkflowModal, ConfirmDeleteModal };
