+++ packages/frinx-workflow-ui/src/pages/executed-workflow-detail/executed-workflow-detail.tsx
@@ -1,5 +1,4 @@
 import React, { ChangeEvent, FC, useEffect, useMemo, useState } from 'react';
-import { TaskModal } from '../../components/modals';
 import {
   Box,
   Button,
@@ -11,7 +10,6 @@ import {
   TabPanels,
   Tabs,
   useToast,
-  useDisclosure,
   Text,
   Progress,
   VStack,
@@ -45,6 +43,7 @@ import DetailsModalHeader from './executed-workflow-detail-header';
 import copyToClipBoard from '../../helpers/copy-to-clipboard';
 import WorkflowDiagram from '../../common/workflow-diagram';
 import { formatDate } from '../../helpers/utils.helpers';
+import ExecutedWorkflowDetailTaskDetail from './executed-workflow-detail-task-detail/executed-workflow-detail-task-detail';
 
 const EXECUTED_WORKFLOW_QUERY = gql`
   query ExecutedWorkflowDetail($nodeId: ID!) {
@@ -241,7 +240,6 @@ type Props = {
 const ExecutedWorkflowDetail: FC<Props> = ({ onExecutedOperation }) => {
   const ctx = useMemo(() => ({ additionalTypenames: ['ExecutedWorkflow'] }), []);
   const { workflowId } = useParams<{ workflowId: string }>();
-  const taskModalDisclosure = useDisclosure();
   const [openedTaskId, setOpenedTaskId] = useState<string | null>(null);
   const [isEscaped, setIsEscaped] = useState(false);
   const [workflowVariables, setWorkflowVariables] = useState<Record<string, string> | null>(null);
@@ -359,16 +357,6 @@ const ExecutedWorkflowDetail: FC<Props> = ({ onExecutedOperation }) => {
     const workflowForm = executedWorkflow.input != null ? JSON.parse(executedWorkflow.input) : {};
     workflowForm[key] = e.target.value;
     setWorkflowVariables((prev) => ({ ...prev, ...workflowForm }));
-  };
-
-  const handleOnOpenTaskModal = (id: string) => {
-    setOpenedTaskId(id);
-    taskModalDisclosure.onOpen();
-  };
-
-  const handleOnCloseTaskModal = () => {
-    setOpenedTaskId(null);
-    taskModalDisclosure.onClose();
   };
 
   const handleCopyToClipborad = (inputText: Record<string, unknown>) => {
@@ -532,14 +520,6 @@ const ExecutedWorkflowDetail: FC<Props> = ({ onExecutedOperation }) => {
 
   return (
     <Container maxWidth={1280}>
-      {openedTaskId != null && (
-        <TaskModal
-          taskId={openedTaskId}
-          executedWorkflow={executedWorkflowDetail.node}
-          isOpen={taskModalDisclosure.isOpen}
-          onClose={handleOnCloseTaskModal}
-        />
-      )}
       <Heading size="xl" marginBottom={10}>
         Details of {executedWorkflow.workflowName} / {executedWorkflow.workflowVersion}
       </Heading>
@@ -566,7 +546,7 @@ const ExecutedWorkflowDetail: FC<Props> = ({ onExecutedOperation }) => {
           <VStack>
             <Progress isIndeterminate />
 
-            <TaskTable tasks={executedWorkflow.tasks} onTaskClick={handleOnOpenTaskModal} />
+            <TaskTable tasks={executedWorkflow.tasks} onTaskClick={setOpenedTaskId} />
           </VStack>
         ) : (
           <Tabs index={tabIndex} onChange={setTabIndex}>
@@ -586,7 +566,15 @@ const ExecutedWorkflowDetail: FC<Props> = ({ onExecutedOperation }) => {
             </TabList>
             <TabPanels>
               <TabPanel>
-                <TaskTable tasks={executedWorkflow.tasks} onTaskClick={handleOnOpenTaskModal} />
+                {openedTaskId == null && <TaskTable tasks={executedWorkflow.tasks} onTaskClick={setOpenedTaskId} />}
+
+                {openedTaskId != null && executedWorkflow.tasks != null && (
+                  <ExecutedWorkflowDetailTaskDetail
+                    executedWorkflow={executedWorkflowDetail.node}
+                    taskId={openedTaskId}
+                    onClose={() => setOpenedTaskId(null)}
+                  />
+                )}
               </TabPanel>
               <TabPanel>
                 <InputOutputTab
